+++ ibizpms-provider/ibizpms-provider-webapi/src/main/java/cn/ibizlab/pms/webapi/dto/TaskTeamDTO.java
@@ -104,7 +104,7 @@ public class TaskTeamDTO extends DTOBase implements Serializable {
      */
     @JSONField(name = "estimate")
     @JsonProperty("estimate")
-    private BigDecimal estimate;
+    private Double estimate;
 
     /**
      * 属性 [ACCOUNT]
@@ -121,7 +121,7 @@ public class TaskTeamDTO extends DTOBase implements Serializable {
      */
     @JSONField(name = "consumed")
     @JsonProperty("consumed")
-    private BigDecimal consumed;
+    private Double consumed;
 
     /**
      * 属性 [ID]
@@ -155,7 +155,7 @@ public class TaskTeamDTO extends DTOBase implements Serializable {
      */
     @JSONField(name = "left")
     @JsonProperty("left")
-    private BigDecimal left;
+    private Double left;
 
 
     /**
@@ -209,7 +209,7 @@ public class TaskTeamDTO extends DTOBase implements Serializable {
     /**
      * 设置 [ESTIMATE]
      */
-    public void setEstimate(BigDecimal  estimate){
+    public void setEstimate(Double  estimate){
         this.estimate = estimate ;
         this.modify("estimate",estimate);
     }
@@ -225,7 +225,7 @@ public class TaskTeamDTO extends DTOBase implements Serializable {
     /**
      * 设置 [CONSUMED]
      */
-    public void setConsumed(BigDecimal  consumed){
+    public void setConsumed(Double  consumed){
         this.consumed = consumed ;
         this.modify("consumed",consumed);
     }
@@ -249,7 +249,7 @@ public class TaskTeamDTO extends DTOBase implements Serializable {
     /**
      * 设置 [LEFT]
      */
-    public void setLeft(BigDecimal  left){
+    public void setLeft(Double  left){
         this.left = left ;
         this.modify("left",left);
     }
